chore(ui): add some unit tests and migrate className utils to index (#3801)
