+++ packages/ui/src/types/primitives/index.ts
@@ -1,1 +1,1 @@
-export { ComponentClassName } from './componentClassName';
+export { ComponentClassName, Modifiers } from './componentClassName';
